Fix issue with M260.9 and M261.9 not supported
RRF does not allow subcommands of M260 and M261 so these have been
renamed to M2600 and M2601 to avoid any sort of conflict.

This commit also fixes an issue with the Shihlin driver where bytes were
not passed through as a vector so could not be counted by the new
`M2600` commands.

Finally, we improve the reliability of Shihlin SL3 driver by setting the
command change flag in the arborState global directly, rather than
setting a local var and updating the global slightly later.

We also lazily load VFD data, so this only occurs when starting the
spindle for the first time. This avoids attempting to read from the VFD
even when the spindle is not required to run.

Signed-off-by: Ben Agricola <ben+git@agrico.la>